project backbone restructurization, code cleanup
