+++ src/common/Loader/index.js
@@ -1,9 +1,7 @@
 import { CircleBorder, Spinner } from "./styled";
 
-const Loader = () => (
+export const Loader = () => (
     <CircleBorder>
         <Spinner />
     </CircleBorder>
 );
-
-export default Loader;
