+++ src/common/MainWrapper/styled.js
@@ -2,6 +2,8 @@ import styled from "styled-components";
 
 export const MainWrapper = styled.main`
     max-width: 1400px;
-    margin: 64px auto 336px;
-    padding: 0 16px;
+    padding:0 16px;
+    margin: 0 auto;
+    display: flex;
+    flex-direction: column;
 `;
