+++ src/features/movies/MovieList/styled.js
@@ -1,26 +1,5 @@
 import styled from "styled-components";
 
-export const Content = styled.section`
-    max-width: 1400px;
-    padding:0 16px;
-    margin: 0 auto;
-    display: flex;
-    flex-direction: column;
-`;
-
-export const Title = styled.h2`
-    font-size: 36px;
-    margin: 0;
-    padding: 56px 0 24px;
-    line-height: 120%;
-    @media (max-width: ${({ theme }) => theme.breakpoint.tablet}px) {
-        font-size:3.51vw;
-    }
-    @media (max-width: ${({ theme }) => theme.breakpoint.mobile}px) {
-        font-size: 18px;
-    }
-`;
-
 export const Wrapper = styled.ul`
     padding: 0;
     display: grid;
