+++ src/features/people/PeopleList/index.js
@@ -2,11 +2,13 @@ import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useParams } from "react-router-dom";
 import { ErrorPage } from "../../../common/ErrorPage";
-import Loader from "../../../common/Loader";
+import { Loader } from "../../../common/Loader";
 import { Pagination } from "../../../common/Pagination";
 import { fetchPeople, selectPeopleFetchStatus, selectPeopleList } from "../peopleSlice";
-import { PersonTile } from "./PersonTile";
-import { Content, Title, Wrapper } from "./styled";
+import { PersonTile } from "../../../common/PersonTile";
+import { Wrapper } from "../../movies/MovieList/styled";
+import { MainWrapper } from "../../../common/MainWrapper/styled";
+import { Header } from "../../../common/Header/styled";
 
 export const PeopleList = () => {
     const dispatch = useDispatch();
@@ -18,26 +20,26 @@ export const PeopleList = () => {
         dispatch(fetchPeople(page));
     }, [dispatch, page]);
 
-    switch (true) {
-        case (fetchStatus === "completed" && people !== undefined):
+    switch (fetchStatus) {
+        case ("completed"):
             return (
-                <Content>
-                    <Title>Popular people</Title>
+                <MainWrapper>
+                    <Header>Popular people</Header>
                     <Wrapper>
                         {people.map((person) =>
                             <PersonTile key={person.id}
                                 name={person.name}
-                                portrait={ person.profile_path ? `https://image.tmdb.org/t/p/w185/${person.profile_path}` : null}
+                                portrait={person.profile_path ? `https://image.tmdb.org/t/p/w185/${person.profile_path}` : null}
                                 id={person.id}
-                                />
+                            />
                         )}
                     </Wrapper>
-                  <Pagination currentPage={page} allPages={500}/>
-                </Content>
+                    <Pagination currentPage={page} allPages={500} />
+                </MainWrapper>
             );
-        case (fetchStatus === "error"):
-            return (<ErrorPage />);
+        case ("error"):
+            return <ErrorPage />
         default:
-            return (<Loader />);
+            return <Loader />
     }
 };
